Update both pages - updated profile and read me
(1) Another program may be editing the same file.  If this is the case,
    be careful not to end up with two different instances of the same
    file when making changes.  Quit, or continue with caution.
(2) An edit session for this file crashed.
    If this is the case, use ":recover" or "vim -r C:/Users/Austin Samona/Deskto
p/Profile/.git/COMMIT_EDITMSG"
    to recover the
(1) Another program may be editing the same file.  If this is the case,
    be careful not to end up with two different instances of the same
    file when making changes.  Quit, or continue with caution.
(2) An edit session for this file crashed.
    If this is the case, use ":recover" or "vim -r C:/Users/Austin Samona/Deskto
p/Profile/.git/COMMIT_EDITMSG"
    to recover the